requirements.txt
certifi==2022.9.14
charset-normalizer==2.1.1
Django==2.2.16
python-dotenv==0.21.0
pytz==2022.2.1
requests==2.28.1
sqlparse==0.4.2
stripe==4.1.0
urllib3==1.26.12